practice 9-1 fixed bug
Problem was in Shop.java class.  Product p1 was initialized not_rated and assumed was that pm.review method would change the product in the ProductManager class, but as that method just returns a product, it did add a review to the productlist, but the average-holder was returned without being stored. long story short: adding 'p1 =' fixed it.